services/face-detection/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
loguru>=0.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

